New ring constructor, GAPSingular function.
Add a new generic constructor for singular rings with orderings.
This should enable us to create nearly arbitrary Singular rings.

Start a function GAPSingular which transforms a singular objects into
a GAP object as far as possible or otherwise returns fail. This can
so far work with Singular strings and small integers, enabling us to do
  Print(GAPSingular(SI_print(ob)));
where ob is an arbitrary Singular object (wrapped for GAP).
Note: Proxy objects are not yet taken care of.