Fix a segfault in SingObj::cleanup()
Multiple cases were actually broken. However, the only code paths in
SingObj::cleanup that ever run are those for bigint and string, and
only the bigint case was affected.

The other cases in SingObj::cleanup are never run. We should get rid
of them by refactoring the code to make this crystal clear.