Move gapwrap() out of class SingObj
The resulting code now is actually simpler and shorter than it was
before. The single purpose of gapwrap() is to take a Singular
interpreter value, an sleftv, and wrap its content in a GAP object. For
this it does not need any of the features of SingObj, so it didn't
really make sense for it to be a member of SingObj.