Fix crash in SI_ringlist
Our heuristic for functions returning ring dependent data is to
either take the ring of any ring dependent input there is; or as
a fallback, the current ring. But when calling SI_ringlist,
the only parameter is a ring, which is of course *not* ring
dependent. And no current ring is set.

The fix is clear: We extend the heuristic to also take rings
occurring as input into account. Moreover, we add a check that
causes us to error out if we really can't figure out any ring,
which is still better than crashing due peeking at a null pointer.