Fix crash when invoking library functions.
We now clear iiRETURNEXPR before invoking iiMake_proc. If we don't do
that, the previous value is still partially in there (some fields are
apparently cleared elsewhere, but not all -- in particular, the type
field remains). If this happens to be a ring, and if moreover the
called procedure runs into an error (and thus does not reset
iiRETURNEXPR), then the Singular interpreter erroneously thinks that
the current call returned a ring. It then tried to dereference
iiRETURNEXPR->data; unfortunately, *that* has been set to NULL.