Use Singular interpreter objects in Func_SI_CallFuncM
This avoids creating tons of copies of objects, and in particular
some pathological cases that could lead to memory usage
exploding through the roof.

Note that this change is incomplete: Func_SI_CallFunc2/3
should also do this, and also Func_SI_CallFuncM for more
than 3 parameters.

Moreover, it might be a good idea to make the handling
of return values a bit more clever: If the return value
is an IHDL, instead of copying it, we could check
if it refers to one of the input objects, and if so,
reuse the GAP object for that.