Mark GAP wrappers for Singular strings as immutable.
Singular strings cannot be modified on the Singular interpreter level.
Hence some Singular (library) code might rely on that, and be confused
if Singular strings are changed after creation.

We can revise this revision in the future if we deem it absolutely
necessary, but for now, and after discussion with Hans, I prefer to go
with the conservative approach.