Print Singular polynomials just like GAP polynomials
This is more user-friendly. It can be reverted to the old way by
setting SI_DEBUG_MODE to true.

That is, almost the old way, as we also changed _SI_ring to set the
shortOut flag to false, to get a more GAP-like output (i.e. "2*x^2"
instead of "2x2")