Do not show whether objects are mutable in their view / print string
For other GAP objects, the mutability status is normally not shown
either. If one truly wants to know, IsMutable() is available. Worse,
when we print "mutable", the user expects to be able to modify the
object, but right now this is not always actually possible...

One drawback is that our test suite looses the checks for "correct"
mutability. However, these checks were really only about checking that
the current behavior does not change; but it is unclear (to me, at
least) whether the current behavior is even *correct* in general.

So, we really should have dedicated tests to mutability for all kinds
of types and operations, which should be consciously written after
deciding what should have which mutability status when.

In any case, the old behavior can be turned back on by setting
SI_DEBUG_MODE to true.