Recreate gh-pages branch using latest GitHubPagesForGAP
The old branch also was based on GitHubPagesForGAP, but on a very old version,
and apparently on one that diverged from the public GitHubPagesForGAP history
(indeed, GitHubPagesForGAP was created in tandem with the SingularInterace
site).

In order to update to the latest GitHubPagesForGAP, the easiest solution
was to create a fresh history.